+++ beemapi/noderpc.py
@@ -11,7 +11,7 @@ import logging
 log = logging.getLogger(__name__)
 
 
-class SteemNodeRPC(GrapheneRPC):
+class NodeRPC(GrapheneRPC):
     """ This class allows to call API methods exposed by the witness node via
         websockets / rpc-json.
 
@@ -27,7 +27,7 @@ class SteemNodeRPC(GrapheneRPC):
     """
 
     def __init__(self, *args, **kwargs):
-        """ Init SteemNodeRPC
+        """ Init NodeRPC
 
             :param str urls: Either a single Websocket/Http URL, or a list of URLs
             :param str user: Username for Authentication
@@ -37,7 +37,7 @@ class SteemNodeRPC(GrapheneRPC):
             :param int timeout: Timeout setting for https nodes (default is 60)
 
         """
-        super(SteemNodeRPC, self).__init__(*args, **kwargs)
+        super(NodeRPC, self).__init__(*args, **kwargs)
         self.next_node_on_empty_reply = False
 
     def set_next_node_on_empty_reply(self, next_node_on_empty_reply=True):
@@ -61,7 +61,7 @@ class SteemNodeRPC(GrapheneRPC):
             doRetry = False
             try:
                 # Forward call to GrapheneWebsocketRPC and catch+evaluate errors
-                reply = super(SteemNodeRPC, self).rpcexec(payload)
+                reply = super(NodeRPC, self).rpcexec(payload)
                 if self.next_node_on_empty_reply and not bool(reply) and self.nodes.working_nodes_count > 1:
                     self._retry_on_next_node("Empty Reply")
                     doRetry = True