+++ tests/beemapi/test_noderpc.py
@@ -12,8 +12,7 @@ import random
 import itertools
 from pprint import pprint
 from beem import Steem
-from beemapi.steemnoderpc import SteemNodeRPC
-from beemapi.websocket import SteemWebsocket
+from beemapi.noderpc import NodeRPC
 from beemapi import exceptions
 from beemapi.exceptions import NumRetriesReached, CallRetriesReached
 from beem.instance import set_shared_steem_instance
@@ -42,7 +41,7 @@ class Testcases(unittest.TestCase):
             keys={"active": wif, "owner": wif, "memo": wif},
             num_retries=10
         )
-        cls.rpc = SteemNodeRPC(urls=cls.nodes_steemit)
+        cls.rpc = NodeRPC(urls=cls.nodes_steemit)
         # from getpass import getpass
         # self.bts.wallet.unlock(getpass())
         set_shared_steem_instance(cls.nodes_steemit)
@@ -85,7 +84,7 @@ class Testcases(unittest.TestCase):
         for node in self.nodes:
             str_list += node + ";"
         str_list = str_list[:-1]
-        rpc = SteemNodeRPC(urls=str_list)
+        rpc = NodeRPC(urls=str_list)
         self.assertIn(rpc.url, self.nodes + self.nodes_steemit)
         rpc.next()
         self.assertIn(rpc.url, self.nodes + self.nodes_steemit)
@@ -95,7 +94,7 @@ class Testcases(unittest.TestCase):
         for node in self.nodes:
             str_list += node + ","
         str_list = str_list[:-1]
-        rpc = SteemNodeRPC(urls=str_list)
+        rpc = NodeRPC(urls=str_list)
         self.assertIn(rpc.url, self.nodes + self.nodes_steemit)
         rpc.next()
         self.assertIn(rpc.url, self.nodes + self.nodes_steemit)
@@ -163,21 +162,21 @@ class Testcases(unittest.TestCase):
         with self.assertRaises(
             NumRetriesReached
         ):
-            SteemNodeRPC(urls="https://wrong.link.com", num_retries=2, timeout=1)
+            NodeRPC(urls="https://wrong.link.com", num_retries=2, timeout=1)
         with self.assertRaises(
             NumRetriesReached
         ):
-            SteemNodeRPC(urls="https://wrong.link.com", num_retries=3, num_retries_call=3, timeout=1)
+            NodeRPC(urls="https://wrong.link.com", num_retries=3, num_retries_call=3, timeout=1)
         nodes = ["https://httpstat.us/500", "https://httpstat.us/501", "https://httpstat.us/502", "https://httpstat.us/503",
                  "https://httpstat.us/505", "https://httpstat.us/511", "https://httpstat.us/520", "https://httpstat.us/522",
                  "https://httpstat.us/524"]
         with self.assertRaises(
             NumRetriesReached
         ):
-            SteemNodeRPC(urls=nodes, num_retries=0, num_retries_call=0, timeout=1)
+            NodeRPC(urls=nodes, num_retries=0, num_retries_call=0, timeout=1)
 
     def test_error_handling(self):
-        rpc = SteemNodeRPC(urls=self.nodes_steemit, num_retries=2, num_retries_call=3)
+        rpc = NodeRPC(urls=self.nodes_steemit, num_retries=2, num_retries_call=3)
         with self.assertRaises(
             exceptions.NoMethodWithName
         ):
@@ -188,7 +187,7 @@ class Testcases(unittest.TestCase):
             rpc.get_accounts("test")
 
     def test_error_handling_appbase(self):
-        rpc = SteemNodeRPC(urls=self.nodes_steemit, num_retries=2, num_retries_call=3)
+        rpc = NodeRPC(urls=self.nodes_steemit, num_retries=2, num_retries_call=3)
         with self.assertRaises(
             exceptions.NoMethodWithName
         ):