+++ tests/beemapi/test_websocket.py
@@ -10,7 +10,7 @@ import random
 import itertools
 from pprint import pprint
 from beem import Steem
-from beemapi.websocket import SteemWebsocket
+from beemapi.websocket import NodeWebsocket
 from beem.instance import set_shared_steem_instance
 from beem.nodelist import NodeList
 # Py3 compatibility
@@ -25,10 +25,10 @@ class Testcases(unittest.TestCase):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         nodelist = NodeList()
-        nodelist.update_nodes(steem_instance=Steem(node=nodelist.get_nodes(normal=True, appbase=True), num_retries=10))
-        stm = Steem(node=nodelist.get_nodes())
+        nodelist.update_nodes(steem_instance=Steem(node=nodelist.get_steem_nodes(), num_retries=10))
+        stm = Steem(node=nodelist.get_steem_nodes())
 
-        self.ws = SteemWebsocket(
+        self.ws = NodeWebsocket(
             urls=stm.rpc.nodes,
             num_retries=10
         )